Update to work with COSA a13cf77
In COSA we dropped the compression of hyperv and applehv
in the OSBuild invocation [1] so we update here for that.

In COSA upstream we also merged in patches for Live ISO
building but that hasn't landed in RPMS yet so we surgically
remove it here until those new stages land.

[1] https://github.com/coreos/coreos-assembler/pull/3982