Update COSA for new platforms support
Since we've added a bunch of new platforms recently in COSA let's
add them here too.

This commit also does some refactor to make it easier to add a new
platform by deriving the list of files to retrieve by iterating
over the keys of the SUPPORTED_PLATFORMS associative array.